API/requirements.txt
flask
flask-cors
python-dotenv
supabase
google-genai
